+++ seantis/people/__init__.py
@@ -6,3 +6,7 @@ _ = MessageFactory('seantis.people')
 
 # duplicated in profiles/default/toolset/toolset.xml
 catalog_id = 'seantis_people_catalog'
+
+import warnings
+from exceptions import UnicodeWarning
+warnings.filterwarnings('error', category=UnicodeWarning)