+++ seantis/people/content/list.py
@@ -92,7 +92,8 @@ class List(Container):
                 if isinstance(filter, LetterFilter):
                     keyword_index = 'first_letter'
 
-                query[keyword_index] = {'query': filter.value}
+                # the catalog cannot deal with unicode filter values
+                query[keyword_index] = {'query': filter.value.encode('utf-8')}
 
         if unrestricted_search:
             return catalog.unrestrictedSearchResults(query)